refactor Canvas
