+++ ui/src/lib/nodes.tsx
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useState, useContext } from "react";
 import { applyNodeChanges, Node } from "reactflow";
 import { RepoContext, RoleType } from "./store";
 import { useStore } from "zustand";
+import { useApolloClient } from "@apollo/client";
 
 const isNodeAddChange = (change) => change.type === "add";
 const isNodeRemoveChange = (change) => change.type === "remove";
@@ -20,13 +21,14 @@ export function resetSelection() {
   return true;
 }
 
-export function useNodesStateSynced(nodeList) {
+export function useNodesStateSynced() {
   const store = useContext(RepoContext);
   if (!store) throw new Error("Missing BearContext.Provider in the tree");
   const addPod = useStore(store, (state) => state.addPod);
   const getPod = useStore(store, (state) => state.getPod);
   const deletePod = useStore(store, (state) => state.deletePod);
   const updatePod = useStore(store, (state) => state.updatePod);
+  const apolloClient = useApolloClient();
   const role = useStore(store, (state) => state.role);
   const ydoc = useStore(store, (state) => state.ydoc);
   const nodesMap = ydoc.getMap<Node>("pods");
@@ -35,76 +37,94 @@ export function useNodesStateSynced(nodeList) {
     (state) => state.provider?.awareness?.clientID
   );
 
-  const [nodes, setNodes] = useState(nodeList);
+  const [nodes, setNodes] = useState<Node[]>([]);
   // const setNodeId = useStore((state) => state.setSelectNode);
   // const selected = useStore((state) => state.selectNode);
 
-  function selectPod(id, selected) {
-    if (selected) {
-      const p = getPod(id)?.parent;
+  const selectPod = useCallback(
+    (id, selected) => {
+      if (selected) {
+        const p = getPod(id)?.parent;
 
-      // if you select a node that has a different parent, clear all previous selections
-      if (parent !== undefined && parent !== p) {
-        selectedPods.clear();
-        setNodes((nds) => nds.map((n) => ({ ...n, selected: false })));
+        // if you select a node that has a different parent, clear all previous selections
+        if (parent !== undefined && parent !== p) {
+          selectedPods.clear();
+          setNodes((nds) => nds.map((n) => ({ ...n, selected: false })));
+        }
+        parent = p;
+        selectedPods.add(id);
+      } else {
+        if (!selectedPods.delete(id)) return;
+        if (selectedPods.size === 0) parent = undefined;
       }
-      parent = p;
-      selectedPods.add(id);
-    } else {
-      if (!selectedPods.delete(id)) return;
-      if (selectedPods.size === 0) parent = undefined;
-    }
-    setNodes((nds) => nds.map((n) => (n.id === id ? { ...n, selected } : n)));
-  }
+      setNodes((nds) => nds.map((n) => (n.id === id ? { ...n, selected } : n)));
+    },
+    [getPod]
+  );
 
-  const onNodesChanges = useCallback((changes) => {
-    const nodes = Array.from(nodesMap.values());
+  const onNodesChange = useCallback(
+    (changes) => {
+      const nodes = Array.from(nodesMap.values());
 
-    const nextNodes = applyNodeChanges(changes, nodes);
+      const nextNodes = applyNodeChanges(changes, nodes);
 
-    // prevent updates from guest users
-    if (role === RoleType.GUEST) {
-      setNodes(nextNodes);
-      return;
-    }
+      // prevent updates from guest users
+      if (role === RoleType.GUEST) {
+        setNodes(nextNodes);
+        return;
+      }
 
-    changes.forEach((change) => {
-      if (!isNodeAddChange(change)) {
-        if (isNodeRemoveChange(change)) {
-          nodesMap.delete(change.id);
-          return;
-        }
-        const node = nextNodes.find((n) => n.id === change.id);
-
-        if (!node) return;
-
-        if (isNodeResetChange(change) || change.type === "select") {
-          selectPod(node.id, change.selected);
-          return;
-        }
-
-        if (change.type === "dimensions" && node.type === "code") {
-          // There is a (seemingly unnecessary) dimension change at the very
-          // beginning of canvas page, which causes dirty status of all
-          // CodeNodes to be set. This is a workaround to prevent that.
-          if (getPod(node.id).width !== node.width) {
-            // only sync width
-            updatePod({
-              id: node.id,
-              data: {
-                width: node.style?.width as number,
-              },
-            });
+      changes.forEach((change) => {
+        if (!isNodeAddChange(change)) {
+          if (isNodeRemoveChange(change)) {
+            nodesMap.delete(change.id);
+            return;
           }
-          return;
-        }
+          const node = nextNodes.find((n) => n.id === change.id);
 
-        if (node) {
-          nodesMap.set(change.id, node);
+          if (!node) return;
+
+          if (isNodeResetChange(change) || change.type === "select") {
+            selectPod(node.id, change.selected);
+            return;
+          }
+
+          if (change.type === "dimensions" && node.type === "code") {
+            // There is a (seemingly unnecessary) dimension change at the very
+            // beginning of canvas page, which causes dirty status of all
+            // CodeNodes to be set. This is a workaround to prevent that.
+            if (getPod(node.id).width !== node.width) {
+              // only sync width
+              updatePod({
+                id: node.id,
+                data: {
+                  width: node.style?.width as number,
+                },
+              });
+            }
+            return;
+          }
+
+          if (node) {
+            nodesMap.set(change.id, node);
+          }
         }
-      }
-    });
-  }, []);
+      });
+    },
+    [getPod, nodesMap, role, selectPod, updatePod]
+  );
+
+  const triggerUpdate = useCallback(() => {
+    setNodes(
+      Array.from(nodesMap.values())
+        .filter(
+          (node) =>
+            !node.data.hasOwnProperty("clientId") ||
+            node.data.clientId === clientId
+        )
+        .sort((a: Node, b: Node) => a.data.level - b.data.level)
+    );
+  }, [clientId, nodesMap]);
 
   useEffect(() => {
     const observer = (YMapEvent) => {
@@ -128,11 +148,14 @@ export function useNodesStateSynced(nodeList) {
         } else if (change.action === "delete") {
           const node = change.oldValue;
           console.log("todelete", node);
-          deletePod(null, { id: node.id, toDelete: [] });
+          deletePod(apolloClient, { id: node.id, toDelete: [] });
         }
       });
 
-      // TOFIX: a node may be shadowed behind its parent, due to the order to render reactflow node, to fix this, comment out the following sorted method, which brings in a large overhead.
+      // TOFIX: a node may be shadowed behind its parent, due to the order to
+      // render reactflow node, to fix this, comment out the following sorted
+      // method, which brings in a large overhead.
+
       setNodes(
         Array.from(nodesMap.values())
           .filter(
@@ -140,7 +163,7 @@ export function useNodesStateSynced(nodeList) {
               !node.data.hasOwnProperty("clientId") ||
               node.data.clientId === clientId
           )
-          .sort((a: Node & { level }, b: Node & { level }) => a.level - b.level)
+          .sort((a: Node, b: Node) => a.data.level - b.data.level)
           .map((node) => ({
             ...node,
             selected: selectedPods.has(node.id),
@@ -158,7 +181,12 @@ export function useNodesStateSynced(nodeList) {
       nodesMap.unobserve(observer);
       resetSelection();
     };
-  }, []);
+  }, [addPod, apolloClient, clientId, deletePod, getPod, nodesMap]);
 
-  return [nodes.filter((n) => n), setNodes, onNodesChanges];
+  return {
+    nodes: nodes.filter((n) => n),
+    onNodesChange,
+    setNodes,
+    triggerUpdate,
+  };
 }
