+++ ui/src/lib/store/repoMetaSlice.tsx
@@ -6,14 +6,6 @@ import { WebsocketProvider } from "y-websocket";
 import { MyState } from ".";
 import { gql } from "@apollo/client";
 
-let serverURL;
-if (window.location.protocol === "http:") {
-  serverURL = `ws://${window.location.host}/socket`;
-} else {
-  serverURL = `wss://${window.location.host}/socket`;
-}
-console.log("yjs server url: ", serverURL);
-
 export interface RepoMetaSlice {
   repoName: string | null;
   repoNameSyncing: boolean;
@@ -38,21 +30,6 @@ export const createRepoMetaSlice: StateCreator<
     set(
       produce((state: MyState) => {
         state.repoId = repoId;
-        state.ydoc = new Doc();
-        // console.log("user reset state setrepo", repoId);
-        if (state.provider) {
-          console.log("emmm, provider exists", state.provider);
-        } else {
-          console.log("connecting yjs socket ..");
-          state.provider = new WebsocketProvider(
-            serverURL,
-            state.repoId,
-            state.ydoc
-          );
-          // max retry time: 10s
-          state.provider.connect();
-          state.provider.maxBackoffTime = 10000;
-        }
       })
     ),
   setRepoName: (name) => {
